Extend some paging-related stuff.
* Book necessary update flags for future support of paged-in ROMType

* Increase BigPageIndex type to 16 bits

+++ src/beeb/include/beeb/type.h
@@ -124,7 +124,7 @@ struct M6502Config;
 
 // See comment for (very similar) CycleCount struct.
 struct BigPageIndex {
-    typedef uint8_t Type;
+    typedef uint16_t Type;
     Type i;
 };
